+++ controllers/contacts.js
@@ -1,0 +1,89 @@
+const Joi = require('joi');
+
+const contacts = require('../models/contacts');
+
+const {HttpError} = require('../helpers');
+
+const addSchema = Joi.object({
+    name: Joi.string().required(),
+    email: Joi.string().required(),
+    phone: Joi.string().required(),
+  });
+
+  const listContacts = async (req, res, next) => {
+    try {
+      const result = await contacts.listContacts()
+      res.json(result)
+    } 
+    catch (error) {
+      next(error)
+    }
+  };
+
+  const getContactById = async (req, res, next) => {
+    try {
+      const {id} = req.params;
+      const result = await contacts.getContactById(id);
+      if(!result) {
+        throw HttpError(404, 'Not found')
+      };
+      res.json(result);
+    } 
+    catch (error) {
+      next(error)
+    };
+  };
+
+  const addContact = async (req, res, next) => {
+    try {
+      const {error} = addSchema.validate(req.body);
+      if(error) {
+        throw HttpError(400, 'missing required name field')
+      };
+      const result = await contacts.addContact(req.body);
+      res.status(201).json(result);
+    } 
+    catch (error) {
+      next(error)
+    };
+  };
+
+  const removeContact = async (req, res, next) => {
+    try {
+      const {id} = req.params;
+      const result = await contacts.removeContact(id);
+      if(!result) {
+        throw HttpError(404, 'Not found')
+      };
+      res.status(200).json({message: "contact deleted"});
+    } 
+    catch (error) {
+      next(error)
+    }
+  };
+
+  const updateContact = async (req, res, next) => {
+    try {
+      const {error} = addSchema.validate(req.body);
+      if(error) {
+        throw HttpError(400, 'missing required name field')
+      };
+      const {id} = req.params;
+      const result = await contacts.updateContact(id, req.body);
+      if(!result) {
+        throw HttpError(404, 'Not found')
+      };
+      res.json(result);
+    } 
+    catch (error) {
+      next(error)
+    };
+  };
+
+  module.exports = {
+    listContacts,
+    getContactById,
+    addContact,
+    removeContact,
+    updateContact,
+  };
