+++ routes/api/contacts.js
@@ -1,86 +1,23 @@
 const express = require('express');
-const Joi = require('joi');
 
-const contacts = require('../../models/contacts');
-
-const {HttpError} = require('../../helpers');
+const {
+  listContacts, 
+  getContactById, 
+  addContact,
+  removeContact,
+  updateContact
+} = require('../../controllers/contacts');
 
 const router = express.Router();
 
-const addSchema = Joi.object({
-  name: Joi.string().required(),
-  email: Joi.string().required(),
-  phone: Joi.string().required(),
-})
+router.get('/', listContacts);
 
-router.get('/', async (req, res, next) => {
-  try {
-    const result = await contacts.listContacts()
-    res.json(result)
-  } 
-  catch (error) {
-    next(error)
-  }
-});
+router.get('/:id', getContactById);
 
-router.get('/:id', async (req, res, next) => {
-  try {
-    const {id} = req.params;
-    const result = await contacts.getContactById(id);
-    if(!result) {
-      throw HttpError(404, 'Not found')
-    };
-    res.json(result);
-  } 
-  catch (error) {
-    next(error)
-  };
-});
+router.post('/', addContact);
 
-router.post('/', async (req, res, next) => {
-  try {
-    const {error} = addSchema.validate(req.body);
-    if(error) {
-      throw HttpError(400, 'missing required name field')
-    };
-    const result = await contacts.addContact(req.body);
-    res.status(201).json(result);
-  } 
-  catch (error) {
-    next(error)
-  };
-});
+router.delete('/:id', removeContact);
 
-router.delete('/:id', async (req, res, next) => {
-  try {
-    const {id} = req.params;
-    const result = await contacts.removeContact(id);
-    if(!result) {
-      throw HttpError(404, 'Not found')
-    };
-    res.status(200).json({message: "contact deleted"});
-  } 
-  catch (error) {
-    next(error)
-  }
-})
-
-router.put('/:id', async (req, res, next) => {
-  try {
-    const {error} = addSchema.validate(req.body);
-    if(error) {
-      throw HttpError(400, 'missing required name field')
-    };
-    const {id} = req.params;
-    const result = await contacts.updateContact(id, req.body);
-    if(!result) {
-      throw HttpError(404, 'Not found')
-    };
-    res.json(result);
-  } 
-  catch (error) {
-    next(error)
-  };
-});
+router.put('/:id', updateContact);
 
 module.exports = router;
